Split prior weights catalog tests into focused modules.
Separates category rules, Laplace helpers, catalog resolution, freighter priors, and solver integration so the suite can grow without another 500-line file.
